Readme tweaks, and move isIP* static methods (#128)
* Update README.md

* move isIP static methods
