+++ lib/socket.js
@@ -24,18 +24,6 @@ module.exports = class UDXSocket extends events.EventEmitter {
     this.userData = null
   }
 
-  static isIPv4 (host) {
-    return ip.isIPv4(host)
-  }
-
-  static isIPv6 (host) {
-    return ip.isIPv6(host)
-  }
-
-  static isIP (host) {
-    return ip.isIP(host)
-  }
-
   get bound () {
     return this._port !== 0
   }
