+++ lib/udx.js
@@ -1,5 +1,6 @@
 const b4a = require('b4a')
 const binding = require('./binding')
+const ip = require('./ip')
 const Socket = require('./socket')
 const Stream = require('./stream')
 const NetworkInterfaces = require('./network-interfaces')
@@ -16,6 +17,18 @@ module.exports = class UDX {
     this._reallocMessage()
 
     binding.udx_napi_init(this._handle, this._buffer)
+  }
+
+  static isIPv4 (host) {
+    return ip.isIPv4(host)
+  }
+
+  static isIPv6 (host) {
+    return ip.isIPv6(host)
+  }
+
+  static isIP (host) {
+    return ip.isIP(host)
   }
 
   _consumeMessage (len) {
